chore: update group post time in user profile
