+++ app/src/main/java/com/example/nocket/constants/Month.kt
@@ -1,0 +1,18 @@
+package com.example.nocket.constants
+
+enum class Month(
+    val displayName: String
+) {
+    JANUARY("January"),
+    FEBRUARY("February"),
+    MARCH("March"),
+    APRIL("April"),
+    MAY("May"),
+    JUNE("June"),
+    JULY("July"),
+    AUGUST("August"),
+    SEPTEMBER("September"),
+    OCTOBER("October"),
+    NOVEMBER("November"),
+    DECEMBER("December")
+}
