+++ app/src/main/java/com/example/nocket/ui/screen/profile/UserProfileScreen.kt
@@ -3,25 +3,17 @@ package com.example.nocket.ui.screen.profile
 import android.os.Build
 import android.util.Log
 import androidx.annotation.RequiresApi
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -34,10 +26,6 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Link
-import androidx.compose.material3.Badge
-import androidx.compose.material3.BadgedBox
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -64,33 +52,21 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import coil3.compose.AsyncImage
+import com.example.nocket.components.empty.EmptyDayItem
+import com.example.nocket.components.grid.PostGridItemWithBadge
 import com.example.nocket.components.topbar.UserProfileTopBar
+import com.example.nocket.constants.Month
 import com.example.nocket.models.Post
 import com.example.nocket.models.auth.AuthState
 import com.example.nocket.models.auth.AuthUser
 import com.example.nocket.ui.screen.postdetail.PostDetailScreen
+import com.example.nocket.utils.calculateDaysOfMonthInYear
+import com.example.nocket.utils.groupPostsByDay
+import com.example.nocket.utils.groupPostsByMonthYear
 import com.example.nocket.utils.mapToUser
 import com.example.nocket.viewmodels.AppwriteViewModel
 import com.example.nocket.viewmodels.AuthViewModel
 import java.time.LocalDate
-import kotlin.math.absoluteValue
-
-enum class Month(
-    val displayName: String
-) {
-    JANUARY("January"),
-    FEBRUARY("February"),
-    MARCH("March"),
-    APRIL("April"),
-    MAY("May"),
-    JUNE("June"),
-    JULY("July"),
-    AUGUST("August"),
-    SEPTEMBER("September"),
-    OCTOBER("October"),
-    NOVEMBER("November"),
-    DECEMBER("December")
-}
 
 data class MonthPosts(
     val month: Month,
@@ -105,17 +81,6 @@ data class DayPostGroup(
     val count: Int get() = posts.size
     val hasMultiplePosts: Boolean get() = posts.size > 1
     val primaryPost: Post? get() = posts.firstOrNull()
-}
-
-fun calculateDaysOfMonthInYear(
-    month: Month,
-    year: Int
-): Int {
-    return when (month) {
-        Month.JANUARY, Month.MARCH, Month.MAY, Month.JULY, Month.AUGUST, Month.OCTOBER, Month.DECEMBER -> 31
-        Month.APRIL, Month.JUNE, Month.SEPTEMBER, Month.NOVEMBER -> 30
-        Month.FEBRUARY -> if ((year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)) 29 else 28
-    }
 }
 
 @RequiresApi(Build.VERSION_CODES.O)
@@ -190,7 +155,7 @@ fun UserProfile(
 
         else -> {
             Log.d("UserProfile", "Displaying profile for user: ${data.username}")
-            Log.d("UserProfile", "Post of user: ${posts.size} posts")
+            Log.d("UserProfile", "Post of user: $posts")
 
             Scaffold(
                 topBar = {
@@ -521,168 +486,3 @@ private fun MonthSection(
     }
 }
 
-private fun groupPostsByDay(posts: List<Post>, daysInMonth: Int): Map<Int, DayPostGroup> {
-    // Group posts by day (mock implementation - in real app, parse createdAt timestamp)
-    val postsByDay = mutableMapOf<Int, MutableList<Post>>()
-
-    posts.forEachIndexed { index, post ->
-        // Mock day assignment - distribute posts across days
-        val day = when {
-            posts.size <= daysInMonth -> index + 1
-            else -> (post.hashCode().absoluteValue % daysInMonth) + 1
-        }
-
-        postsByDay.getOrPut(day) { mutableListOf() }.add(post)
-    }
-
-    return postsByDay.mapValues { (day, posts) ->
-        DayPostGroup(dayNumber = day, posts = posts)
-    }
-}
-
-@Composable
-private fun PostGridItemWithBadge(
-    dayPostGroup: DayPostGroup,
-    onClick: () -> Unit
-) {
-    BadgedBox(
-        badge = {
-            if (dayPostGroup.hasMultiplePosts) {
-                Badge(
-                    containerColor = Color(0xFFFFD700),
-                    contentColor = Color.Black,
-                    modifier = Modifier.offset(x = (-4).dp, y = 4.dp)
-                ) {
-                    Text(
-                        text = "${dayPostGroup.count}",
-                        style = MaterialTheme.typography.labelSmall,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-            }
-        }
-    ) {
-        Card(
-            modifier = Modifier
-                .aspectRatio(1f)
-                .clickable { onClick() },
-            shape = RoundedCornerShape(12.dp),
-            elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-        ) {
-            Box {
-                // Main post image
-                AsyncImage(
-                    model = dayPostGroup.primaryPost?.thumbnailUrl,
-                    contentDescription = "Post image",
-                    modifier = Modifier.fillMaxSize(),
-                    contentScale = ContentScale.Crop
-                )
-
-                // Blur overlay for multiple posts
-                if (dayPostGroup.hasMultiplePosts) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .background(
-                                brush = Brush.verticalGradient(
-                                    colors = listOf(
-                                        Color.Transparent,
-                                        Color.Black.copy(alpha = 0.3f)
-                                    )
-                                )
-                            )
-                    )
-
-                    // Stack indicator
-                    Box(
-                        modifier = Modifier
-                            .align(Alignment.BottomEnd)
-                            .padding(4.dp)
-                            .background(
-                                color = Color.Black.copy(alpha = 0.7f),
-                                shape = RoundedCornerShape(8.dp)
-                            )
-                            .padding(horizontal = 6.dp, vertical = 2.dp)
-                    ) {
-                        Text(
-                            text = "+${dayPostGroup.count - 1}",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = Color.White,
-                            fontWeight = FontWeight.Medium
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun PostGridItem(
-    post: Post,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .aspectRatio(1f)
-            .clickable { onClick() },
-        shape = RoundedCornerShape(12.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-    ) {
-        AsyncImage(
-            model = post.thumbnailUrl,
-            contentDescription = "Post image",
-            modifier = Modifier.fillMaxSize(),
-            contentScale = ContentScale.Crop
-        )
-    }
-}
-
-@Composable
-private fun EmptyDayItem() {
-    Box(
-        modifier = Modifier
-            .aspectRatio(1f)
-            .background(
-                color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.2f),
-                shape = RoundedCornerShape(12.dp)
-            )
-            .border(
-                width = 1.dp,
-                color = MaterialTheme.colorScheme.outline.copy(alpha = 0.1f),
-                shape = RoundedCornerShape(12.dp)
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        // Subtle dot indicator
-        Box(
-            modifier = Modifier
-                .size(4.dp)
-                .background(
-                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.3f),
-                    shape = CircleShape
-                )
-        )
-    }
-}
-
-// Helper function to group posts by month and year
-@RequiresApi(Build.VERSION_CODES.O)
-private fun groupPostsByMonthYear(posts: List<Post>): List<MonthPosts> {
-    val currentDate = LocalDate.now()
-
-    // Create a more realistic distribution that demonstrates badge functionality
-    // Some posts will be on the same day to show the badge feature
-    return listOf(
-        MonthPosts(
-            month = Month.AUGUST,
-            year = 2025,
-            posts = posts // All posts in August to show grouping
-        ),
-        MonthPosts(
-            month = Month.JULY,
-            year = 2025,
-            posts = posts.take(2) // Some posts in July
-        )
-    ).filter { it.posts.isNotEmpty() }
-}
